+++ client/src/client/logstore.rs
@@ -1,0 +1,16 @@
+use super::*;
+
+mod create_logstore;
+pub use create_logstore::*;
+
+mod update_logstore;
+pub use update_logstore::*;
+
+mod delete_logstore;
+pub use delete_logstore::*;
+
+mod get_logstore;
+pub use get_logstore::*;
+
+mod list_logstores;
+pub use list_logstores::*;